feat: route based yat instead of query param
