+++ src/context/ModalProvider/ModalProvider.tsx
@@ -14,7 +14,6 @@ import { ReceiveModal } from 'components/Modals/Receive/Receive'
 import { SendModal } from 'components/Modals/Send/Send'
 import { SettingsModal } from 'components/Modals/Settings/Settings'
 import { AddAccountModal } from 'pages/Accounts/AddAccountModal'
-import { YatModal } from 'pages/Yat/YatModal'
 
 import { ModalContext } from './ModalContext'
 
@@ -36,7 +35,6 @@ const MODALS = {
   nativeOnboard: NativeOnboarding,
   nft: NftModal,
   feedbackSupport: FeedbackAndSupport,
-  yat: YatModal,
 }
 
 // state
